test: storage and proposal similar to console
