+++ src/tests/specs/console/console.cdn.spec.ts
@@ -1,8 +1,4 @@
-import type {
-	_SERVICE as ConsoleActor,
-	HttpRequest,
-	ProposalType
-} from '$declarations/console/console.did';
+import type { _SERVICE as ConsoleActor } from '$declarations/console/console.did';
 import { idlFactory as idlFactorConsole } from '$declarations/console/console.factory.did';
 import { AnonymousIdentity } from '@dfinity/agent';
 import { Ed25519KeyIdentity } from '@dfinity/identity';
@@ -11,21 +7,16 @@ import {
 	arrayBufferToUint8Array,
 	assertNonNullish,
 	fromNullable,
-	toNullable,
-	uint8ArrayToHexString
+	toNullable
 } from '@dfinity/utils';
 import { PocketIc, type Actor } from '@hadronous/pic';
 import { beforeAll, describe, expect, inject } from 'vitest';
 import { CONTROLLER_ERROR_MSG } from '../../constants/console-tests.constants';
-import { mockBlob, mockHtml } from '../../mocks/storage.mocks';
 import {
 	testControlledCdnMethods,
 	testNotAllowedCdnMethods
 } from '../../utils/cdn-assertions-tests.utils';
-import { uploadFile } from '../../utils/cdn-tests.utils';
-import { assertCertification } from '../../utils/certification-tests.utils';
 import { sha256ToBase64String } from '../../utils/crypto-tests.utils';
-import { tick } from '../../utils/pic-tests.utils';
 import { CONSOLE_WASM_PATH } from '../../utils/setup-tests.utils';
 
 describe('Console > Cdn', () => {
@@ -81,377 +72,12 @@ describe('Console > Cdn', () => {
 			actor.setIdentity(controller);
 		});
 
-		testControlledCdnMethods({ actor: () => actor });
-
-		describe.each([
-			{
-				proposal_type: {
-					AssetsUpgrade: {
-						clear_existing_assets: toNullable()
-					}
-				} as ProposalType,
-				collection: '#dapp',
-				full_path: '/hello.html',
-				expected_proposal_id: 1n
-			},
-			{
-				proposal_type: {
-					SegmentsDeployment: {
-						mission_control_version: [],
-						orbiter: [],
-						satellite_version: ['0.0.18']
-					}
-				} as ProposalType,
-				collection: '#releases',
-				full_path: '/releases/satellite-v0.0.18.wasm.gz',
-				expected_proposal_id: 2n
-			}
-		])(
-			'Proposal, upload and serve',
-			({ proposal_type, collection, full_path, expected_proposal_id }) => {
-				let sha256: [] | [Uint8Array | number[]];
-				let proposalId: bigint;
-
-				it('should init a proposal', async () => {
-					const { init_proposal } = actor;
-
-					const [id, proposal] = await init_proposal(proposal_type);
-
-					proposalId = id;
-
-					expect(proposalId).toEqual(expected_proposal_id);
-
-					expect(proposal.status).toEqual({ Initialized: null });
-					expect(fromNullable(proposal.sha256)).toBeUndefined();
-					expect(fromNullable(proposal.executed_at)).toBeUndefined();
-					expect(proposal.owner.toText()).toEqual(controller.getPrincipal().toText());
-					expect(proposal.proposal_type).toEqual(proposal_type);
-					expect(proposal.created_at).not.toBeUndefined();
-					expect(proposal.created_at).toBeGreaterThan(0n);
-					expect(proposal.updated_at).not.toBeUndefined();
-					expect(proposal.updated_at).toBeGreaterThan(0n);
-					expect(fromNullable(proposal.version) ?? 0n).toBeGreaterThan(0n);
-				});
-
-				it('should fail at uploading asset for unknown proposal', async () => {
-					const { init_asset_upload } = actor;
-
-					const unknownProposalId = proposalId + 1n;
-
-					await expect(
-						init_asset_upload(
-							{
-								collection,
-								description: toNullable(),
-								encoding_type: [],
-								full_path,
-								name: 'hello.html',
-								token: toNullable()
-							},
-							unknownProposalId
-						)
-					).rejects.toThrow(`No proposal found for ${unknownProposalId}`);
-				});
-
-				it('should upload asset', async () => {
-					const { http_request, commit_asset_upload, upload_asset_chunk, init_asset_upload } =
-						actor;
-
-					const file = await init_asset_upload(
-						{
-							collection,
-							description: toNullable(),
-							encoding_type: [],
-							full_path,
-							name: 'hello.html',
-							token: toNullable()
-						},
-						proposalId
-					);
-
-					const chunk = await upload_asset_chunk({
-						batch_id: file.batch_id,
-						content: arrayBufferToUint8Array(await mockBlob.arrayBuffer()),
-						order_id: [0n]
-					});
-
-					await commit_asset_upload({
-						batch_id: file.batch_id,
-						chunk_ids: [chunk.chunk_id],
-						headers: []
-					});
-
-					const { status_code } = await http_request({
-						body: [],
-						certificate_version: toNullable(),
-						headers: [],
-						method: 'GET',
-						url: full_path
-					});
-
-					expect(status_code).toBe(404);
-				});
-
-				it('should fail at submitting an unknown proposal', async () => {
-					const { submit_proposal } = actor;
-
-					const unknownProposalId = proposalId + 1n;
-
-					await expect(submit_proposal(unknownProposalId)).rejects.toThrow(
-						'juno.error.proposals.cannot_submit'
-					);
-				});
-
-				it('should submit proposal', async () => {
-					const { submit_proposal } = actor;
-
-					// Advance time for updated_at
-					await pic.advanceTime(100);
-
-					const [_, proposal] = await submit_proposal(proposalId);
-
-					expect(proposal.status).toEqual({ Open: null });
-					expect(sha256ToBase64String(fromNullable(proposal.sha256) ?? [])).not.toBeUndefined();
-					expect(fromNullable(proposal.executed_at)).toBeUndefined();
-					expect(proposal.owner.toText()).toEqual(controller.getPrincipal().toText());
-					expect(proposal.proposal_type).toEqual(proposal_type);
-					expect(proposal.created_at).not.toBeUndefined();
-					expect(proposal.created_at).toBeGreaterThan(0n);
-					expect(proposal.updated_at).not.toBeUndefined();
-					expect(proposal.updated_at).toBeGreaterThan(0n);
-					expect(proposal.updated_at).toBeGreaterThan(proposal.created_at);
-					expect(fromNullable(proposal.version) ?? 0n).toEqual(2n);
-
-					// eslint-disable-next-line prefer-destructuring
-					sha256 = proposal.sha256;
-				});
-
-				it('should still not serve asset', async () => {
-					const { http_request } = actor;
-
-					const { status_code } = await http_request({
-						body: [],
-						certificate_version: toNullable(),
-						headers: [],
-						method: 'GET',
-						url: full_path
-					});
-
-					expect(status_code).toBe(404);
-				});
-
-				it('should fail at submitting a proposal if already open', async () => {
-					const { submit_proposal } = actor;
-
-					await expect(submit_proposal(proposalId)).rejects.toThrow(
-						'juno.error.proposals.cannot_submit_invalid_status (Open)'
-					);
-				});
-
-				it('should fail at committing a proposal if unknown', async () => {
-					const { commit_proposal } = actor;
-
-					const unknownProposalId = proposalId + 1n;
-
-					await expect(
-						commit_proposal({
-							sha256: Array.from({ length: 32 }).map((_, i) => i),
-							proposal_id: proposalId + 1n
-						})
-					).rejects.toThrow(`juno.error.proposals.cannot_commit (${unknownProposalId})`);
-				});
-
-				it('should fail at committing a proposal with incorrect sha256', async () => {
-					const { commit_proposal } = actor;
-
-					const sha256 = Array.from({ length: 32 }).map((_, i) => i);
-
-					await expect(
-						commit_proposal({
-							sha256,
-							proposal_id: proposalId
-						})
-					).rejects.toThrow(`juno.error.proposals.invalid_hash (${uint8ArrayToHexString(sha256)})`);
-				});
-
-				it('should not update proposal status', async () => {
-					const { get_proposal } = actor;
-
-					const proposal = fromNullable(await get_proposal(proposalId));
-
-					assertNonNullish(proposal);
-
-					expect(proposal.status).toEqual({ Open: null });
-				});
-
-				it('should commit proposal', async () => {
-					const { commit_proposal } = actor;
-
-					await expect(
-						commit_proposal({
-							sha256: fromNullable(sha256)!,
-							proposal_id: proposalId
-						})
-					).resolves.not.toThrow();
-				});
-
-				it('should have updated proposal to executed', async () => {
-					const { get_proposal } = actor;
-
-					const proposal = fromNullable(await get_proposal(proposalId));
-
-					assertNonNullish(proposal);
-
-					expect(proposal.status).toEqual({ Executed: null });
-				});
-
-				it('should serve asset', async () => {
-					const { http_request } = actor;
-
-					await tick(pic);
-
-					const request: HttpRequest = {
-						body: [],
-						certificate_version: toNullable(2),
-						headers: [],
-						method: 'GET',
-						url: full_path
-					};
-
-					const response = await http_request(request);
-
-					const { status_code, headers, body } = response;
-
-					expect(status_code).toBe(200);
-
-					const rest = headers.filter(([header, _]) => header !== 'IC-Certificate');
-
-					/* eslint-disable no-useless-escape */
-					expect(rest).toEqual([
-						['accept-ranges', 'bytes'],
-						['etag', '"03ee66f1452916b4f91a504c1e9babfa201b6d64c26a82b2cf03c3ed49d91585"'],
-						['X-Content-Type-Options', 'nosniff'],
-						['Strict-Transport-Security', 'max-age=31536000 ; includeSubDomains'],
-						['Referrer-Policy', 'same-origin'],
-						['X-Frame-Options', 'DENY'],
-						['Cache-Control', 'no-cache'],
-						[
-							'IC-CertificateExpression',
-							'default_certification(ValidationArgs{certification:Certification{no_request_certification:Empty{},response_certification:ResponseCertification{certified_response_headers:ResponseHeaderList{headers:[\"accept-ranges\",\"etag\",\"X-Content-Type-Options\",\"Strict-Transport-Security\",\"Referrer-Policy\",\"X-Frame-Options\",\"Cache-Control\"]}}}})'
-						]
-					]);
-					/* eslint-enable no-useless-escape */
-
-					await assertCertification({
-						canisterId,
-						pic,
-						request,
-						response,
-						currentDate
-					});
-
-					const decoder = new TextDecoder();
-
-					expect(decoder.decode(body as Uint8Array<ArrayBufferLike>)).toEqual(mockHtml);
-				});
-
-				it('should list assets', async () => {
-					const { list_assets } = actor;
-
-					const assets = await list_assets(collection, {
-						matcher: [],
-						order: [],
-						owner: [],
-						paginate: []
-					});
-
-					for (const [_, { version }] of assets.items) {
-						expect(fromNullable(version)).not.toBeUndefined();
-					}
-
-					expect(assets.items.find(([fullPath]) => fullPath === full_path)).not.toBeUndefined();
-				});
-			}
-		);
-
-		it('should clear assets with proposal', async () => {
-			const { http_request, init_proposal, commit_proposal, submit_proposal } = actor;
-
-			const [proposalId, __] = await init_proposal({
-				AssetsUpgrade: {
-					clear_existing_assets: toNullable(true)
-				}
-			});
-
-			await uploadFile({ proposalId, actor });
-
-			const [_, proposal] = await submit_proposal(proposalId);
-
-			await commit_proposal({
-				sha256: fromNullable(proposal.sha256)!,
-				proposal_id: proposalId
-			});
-
-			const { status_code } = await http_request({
-				body: [],
-				certificate_version: toNullable(),
-				headers: [],
-				method: 'GET',
-				url: '/hello.html'
-			});
-
-			expect(status_code).toEqual(404);
-
-			const { status_code: status_code_200 } = await http_request({
-				body: [],
-				certificate_version: toNullable(),
-				headers: [],
-				method: 'GET',
-				url: '/hello3.html'
-			});
-
-			expect(status_code_200).toEqual(200);
-		});
-
-		it('should still serve asset after #dApp has been cleared', async () => {
-			const { http_request } = actor;
-
-			const { status_code } = await http_request({
-				body: [],
-				certificate_version: toNullable(),
-				headers: [],
-				method: 'GET',
-				url: '/releases/satellite-v0.0.18.wasm.gz'
-			});
-
-			expect(status_code).toBe(200);
-		});
-
-		it('should mark proposal as failed', async () => {
-			const { init_proposal, commit_proposal, submit_proposal, get_proposal } = actor;
-
-			const [proposalId, __] = await init_proposal({
-				AssetsUpgrade: {
-					clear_existing_assets: toNullable(false)
-				}
-			});
-
-			// We do not upload any chunks and commit batch to make the proposal fail.
-
-			const [_, { sha256 }] = await submit_proposal(proposalId);
-
-			await expect(
-				commit_proposal({
-					sha256: fromNullable(sha256)!,
-					proposal_id: proposalId
-				})
-			).rejects.toThrow(`juno.error.proposals.empty_assets (${proposalId})`);
-
-			const proposal = fromNullable(await get_proposal(proposalId));
-
-			assertNonNullish(proposal);
-
-			expect(proposal.status).toEqual({ Failed: null });
+		testControlledCdnMethods({
+			actor: () => actor,
+			currentDate,
+			canisterId: () => canisterId,
+			controller: () => controller,
+			pic: () => pic
 		});
 
 		describe('Releases assertions', () => {
